refactor: decompose 4 more monolith files into submodules (issue #2486 sprint 11)
- urdf_generator.py (787 → 272 LOC) → _link_generation, _joint_generation,
  _geometry_helpers, _inertia_helpers, _xml_builder
- collision_generator.py (778 → 220 LOC) → _cg_types, _cg_convex_hull,
  _cg_primitive_fitting, _cg_decimation, _cg_metrics
- motion_capture.py (769 → 25 LOC) → _mocap_data, _mocap_loader, _mocap_processor,
  _mocap_retargeting, _mocap_validator
- interfaces.py (763 → 88 LOC) → _simulation_interface, _dynamics_interface,
  _recorder_interface

All public symbols remain importable from original module paths.
