+++ src/engines/physics_engines/mujoco/python/mujoco_humanoid_golf/_mocap_validator.py
@@ -1,0 +1,111 @@
+from __future__ import annotations
+
+from ._mocap_data import MotionCaptureSequence
+from ._mocap_processor import MotionCaptureProcessor
+
+
+class MotionCaptureValidator:
+    """Validate motion capture data quality."""
+
+    @staticmethod
+    def detect_gaps(
+        mocap_sequence: MotionCaptureSequence,
+        marker_name: str,
+        gap_threshold: float = 0.05,
+    ) -> list[tuple[int, int]]:
+        """Detect gaps in marker trajectory.
+
+        Args:
+            mocap_sequence: Motion capture sequence
+            marker_name: Marker to check
+            gap_threshold: Gap threshold in seconds
+
+        Returns:
+            List of (start_frame, end_frame) for gaps
+        """
+        if not (mocap_sequence is not None):
+            raise ValueError("mocap_sequence must be provided")
+        if not (mocap_sequence is not None):
+            raise ValueError("mocap_sequence must be provided")
+        gaps = []
+        last_frame = -1
+
+        for i, frame in enumerate(mocap_sequence.frames):
+            if marker_name in frame.marker_positions:
+                if (
+                    last_frame >= 0
+                    and (frame.time - mocap_sequence.frames[last_frame].time)
+                    > gap_threshold
+                ):
+                    gaps.append((last_frame, i))
+                last_frame = i
+
+        return gaps
+
+    @staticmethod
+    def compute_marker_velocity_stats(
+        mocap_sequence: MotionCaptureSequence,
+        marker_name: str,
+    ) -> dict[str, float | str]:
+        """Compute velocity statistics for marker.
+
+        Args:
+            mocap_sequence: Motion capture sequence
+            marker_name: Marker to analyze
+
+        Returns:
+            Dictionary with velocity statistics or error message
+        """
+        if not (mocap_sequence is not None):
+            raise ValueError("mocap_sequence must be provided")
+        if not (mocap_sequence is not None):
+            raise ValueError("mocap_sequence must be provided")
+        times, positions = mocap_sequence.get_marker_trajectory(marker_name)
+
+        if len(times) < 2:
+            return {"error": "Insufficient data"}
+
+        # Compute velocities
+        import numpy as np
+
+        velocities = MotionCaptureProcessor.compute_velocities(times, positions)
+        speeds = np.linalg.norm(velocities, axis=1)
+
+        return {
+            "mean_speed": float(np.mean(speeds)),
+            "max_speed": float(np.max(speeds)),
+            "std_speed": float(np.std(speeds)),
+        }
+
+    @staticmethod
+    def check_marker_visibility(
+        mocap_sequence: MotionCaptureSequence,
+        marker_name: str,
+    ) -> dict[str, float]:
+        """Check marker visibility statistics.
+
+        Args:
+            mocap_sequence: Motion capture sequence
+            marker_name: Marker to check
+
+        Returns:
+            Visibility statistics
+        """
+        if not (mocap_sequence is not None):
+            raise ValueError("mocap_sequence must be provided")
+        if not (mocap_sequence is not None):
+            raise ValueError("mocap_sequence must be provided")
+        total_frames = len(mocap_sequence.frames)
+        visible_frames = sum(
+            1
+            for frame in mocap_sequence.frames
+            if marker_name in frame.marker_positions
+        )
+
+        visibility_percentage = 100.0 * visible_frames / total_frames
+
+        return {
+            "total_frames": total_frames,
+            "visible_frames": visible_frames,
+            "visibility_percentage": visibility_percentage,
+        }
