+++ src/shared/python/humanoid_character_builder/mesh/_cg_metrics.py
@@ -1,0 +1,60 @@
+from __future__ import annotations
+
+from typing import Any
+
+
+def count_triangles(components: list[Any]) -> int:
+    if not (components is not None):
+        raise ValueError("components must be provided")
+    if not (components is not None):
+        raise ValueError("components must be provided")
+    total = 0
+    for comp in components:
+        if hasattr(comp, "faces"):
+            total += len(comp.faces)
+    return total
+
+
+def compute_volume_preservation(
+    original: Any,
+    components: list[Any],
+    original_volume: float,
+) -> float:
+    if not (components is not None):
+        raise ValueError("components must be provided")
+    if not (components is not None):
+        raise ValueError("components must be provided")
+    if original_volume <= 0:
+        return 1.0
+
+    total_volume = 0.0
+    for comp in components:
+        if hasattr(comp, "volume"):
+            total_volume += comp.volume
+
+    return min(total_volume / original_volume, 1.0)
+
+
+def compute_hausdorff_distance(
+    original: Any,
+    components: list[Any],
+) -> float:
+    try:
+        original_points = original.sample(1000)
+        max_dist = 0.0
+
+        for comp in components:
+            if hasattr(comp, "nearest"):
+                _, distances, _ = comp.nearest.on_surface(original_points)
+                max_dist = max(max_dist, distances.max())
+
+        return float(max_dist)
+    except (ValueError, ZeroDivisionError, OverflowError, TypeError):
+        return float("inf")
+
+
+__all__ = [
+    "compute_hausdorff_distance",
+    "compute_volume_preservation",
+    "count_triangles",
+]
