+++ src/shared/python/humanoid_character_builder/mesh/_cg_types.py
@@ -1,0 +1,68 @@
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from enum import Enum, auto
+from typing import Any
+
+
+class SimplificationMethod(Enum):
+    AUTO = auto()
+    VHACD = auto()
+    PRIMITIVES = auto()
+    DECIMATION = auto()
+    CONVEX_HULL = auto()
+    HYBRID = auto()
+
+
+class ComplexityLevel(Enum):
+    MINIMAL = auto()
+    BALANCED = auto()
+    ACCURATE = auto()
+
+
+@dataclass
+class PrimitiveFit:
+    primitive_type: str
+    center: tuple[float, float, float]
+    dimensions: tuple[float, ...]
+    rotation: tuple[float, float, float, float]
+    volume_ratio: float
+    error_metric: float
+
+
+@dataclass
+class CollisionGeometryResult:
+    success: bool
+    method_used: SimplificationMethod
+    components: list[Any]
+    original_triangles: int
+    final_triangles: int
+    reduction_ratio: float
+    volume_preservation: float
+    hausdorff_distance: float
+    primitive_fits: list[PrimitiveFit] = field(default_factory=list)
+    warnings: list[str] = field(default_factory=list)
+    errors: list[str] = field(default_factory=list)
+
+
+@dataclass
+class VHACDParameters:
+    max_hulls: int = 16
+    max_vertices_per_hull: int = 64
+    resolution: int = 100000
+    concavity: float = 0.001
+    plane_downsampling: int = 4
+    hull_downsampling: int = 4
+    alpha: float = 0.05
+    beta: float = 0.05
+    mode: int = 0
+    min_volume_per_hull: float = 0.0001
+
+
+__all__ = [
+    "CollisionGeometryResult",
+    "ComplexityLevel",
+    "PrimitiveFit",
+    "SimplificationMethod",
+    "VHACDParameters",
+]
